+++ src/cli.ts
@@ -47,7 +47,7 @@ import {LogPrefixes, makeConsoleLogger} from "./logger"
 const OutputStdOut = '-'
 
 interface CommandOptions {
-  esbuildWorkingDir: string,
+  buildWorkingDir: string,
   gatherLicenseTexts: boolean
   outputReproducible: boolean
   specVersion: SpecVersion
@@ -57,17 +57,25 @@ interface CommandOptions {
   verbose: number
 }
 
-function makeCommand(process_: NodeJS.Process): Command {
+function makeCommand(): Command {
   return new Command(
     /* auto-set the name */
   ).description(
     'Create CycloneDX Software Bill of Materials (SBOM) from esbuild-compatible metafile.'
   ).addOption(
+    /* deprecated
+       missed the <dir> value in the original implementation - so this was never working as intended
+     */
     new Option(
       '--ewd, --esbuild-working-dir',
+      'DEPRECATED'
+    ).hideHelp()
+  ).addOption(
+    new Option(
+      '--bwd, --build-working-dir <dir>',
       'Working dir used in the build process.'
     ).default(
-      process_.cwd(),
+      '.',
       'current working dir'
     )
   ).addOption(
@@ -152,7 +160,7 @@ function makeCommand(process_: NodeJS.Process): Command {
 export async function run(process_: NodeJS.Process): Promise<number> {
   process_.title = 'cyclonedx-esbuild' /* eslint-disable-line  no-param-reassign -- ack */
 
-  const program = makeCommand(process_)
+  const program = makeCommand()
   program.parse(process_.argv)
 
   const options: CommandOptions = program.opts()
@@ -187,7 +195,7 @@ export async function run(process_: NodeJS.Process): Promise<number> {
   const bom = bomBuilder.fromMetafile(
     /* eslint-disable-next-line  @typescript-eslint/no-unsafe-type-assertion -- ack */
     loadJsonFile(metafile) as esbuild.Metafile,
-    options.esbuildWorkingDir,
+    resolve(process_.cwd(), options.buildWorkingDir),
     options.gatherLicenseTexts,
     options.outputReproducible,
     logger)